refactor: 자신의 위치를 보내는 코드 _app으로 이동
